+++ creatorsapi_python_sdk/api_client.py
@@ -107,7 +107,7 @@ class ApiClient:
             self.default_headers[header_name] = header_value
         self.cookie = cookie
         # Set default User-Agent.
-        self.user_agent = 'creatorsapi-python-sdk/1.2.0'
+        self.user_agent = 'creatorsapi-python-sdk/1.3.0'
         self.client_side_validation = configuration.client_side_validation
         
         # OAuth2 properties